refactored TrackCard to be usable on PlayMixtapePage and MixtapeDetailPage - make it flippable to show description on back
